Listing the events done

+++ src/main/java/simon/remy/ensisa/controller/EventRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "event", path = "event")
-public interface EventRepository extends CrudRepository<Event, Long>{
+public interface EventRepository extends CrudRepository<Event, Integer>{
 
 	List<Event> findByTitle(String title);
 }
